🗺️ Atlas: [architectural change] Extract circuit breaker module
Extracted the inline `circuit_breaker` module out of the large `src/run/swebench.rs` file into its own dedicated module file `src/run/circuit_breaker.rs`. This reduces the file size of `swebench.rs` and improves code structure by decoupling the breaker logic from the sweep runner.
